# Improve statistics of iptsd-perf